Allow implementing different WIT worlds in wasip2 target
As mentioned in [this comment](https://github.com/tinygo-org/tinygo/issues/4843#issuecomment-2988622249), we are also interested in Wasm components with custom worlds (for [Grafbase extensions](https://grafbase.com/docs/gateway/extensions)).

I have succesfully compiled and used examples based on the changes in this branch. They are mostly from @daedric's patch in the issue.

So this PR is meant as a starting point for discussions, a working branch for anyone else looking at trying this, and to discuss how best to test it (I've made attempts, but they added up to a lot of code, and I wasn't too happy with them).

I understand it's work to review and merge something like this, and even more to provide guidance on a proper test setup, so while feedback would be much appreciated, I understand if it's not a priority.

fixes #4843